fix(users): update imports after move
