+++ frontend/src/app/users/pages/users.page.scss
@@ -1,65 +1,72 @@
-/* src/app/users/users.page.scss */
-:host {
-  display: block;
+.users-wrapper {
   background: var(--ion-background-color);
+  padding: 2rem 1rem 3rem;
 }
 
-.search-card {
-  margin-bottom: 1rem;
-  padding: 0.75rem;
-  background: var(--ion-color-light);
-  border-radius: var(--ion-card-border-radius);
-  box-shadow: var(--ion-card-box-shadow);
+.users-container {
+  max-width: 1100px;
+  padding: 0 1rem;
+  margin: 0 auto;
+  display: flex;
+  flex-direction: column;
+  gap: 2.5rem;
 }
 
-.table-container {
-  background: var(--ion-color-light);
-  border-radius: 12px;
-  box-shadow: var(--ion-card-box-shadow);
-  overflow-x: auto;
-}
+/* ===============================
+   Cabeçalho
+================================== */
+.header-section {
+  text-align: center;
 
-.user-table {
-  width: 100%;
-  border-collapse: collapse;
-  font-family: var(--ion-font-family);
-
-  th,
-  td {
-    padding: 0.75rem 1rem;
-    font-size: var(--ion-font-size-md);
-    color: var(--ion-text-color);
+  h1 {
+    font-size: 1.75rem;
+    font-weight: 700;
+    color: #1f2937;
+    margin-bottom: 0.5rem;
   }
 
-  thead th {
-    background: var(--ion-color-primary);
-    color: var(--ion-color-primary-contrast);
+  p {
+    font-size: 0.95rem;
+    color: #6b7280;
+  }
+}
+
+
+/* ===============================
+   Barra de ações
+================================== */
+.actions-bar {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 1rem;
+  justify-content: space-between;
+  align-items: center;
+
+  .user-search {
+    flex: 1;
+    padding: 0;
+    --background: white;
+    --border-radius: 12px;
+    border-radius: 12px;
+    box-shadow: 0 1px 2px rgba(0, 0, 0, 0.05);
+  }
+
+  ion-button {
+    height: 44px;
     font-weight: 600;
-    border-bottom: 2px solid var(--ion-color-secondary);
+    --background: var(--ion-color-primary);
+    --border-radius: 12px;
+    --color: var(--ion-color-primary-contrast);
   }
+}
 
-  tbody tr {
-    border-bottom: 1px solid var(--ion-color-secondary-shade);
-    transition: background-color 0.2s;
 
-    &:hover {
-      background: var(--ion-color-secondary-tint);
-    }
-  }
-
-  .actions-header {
-    text-align: center;
-  }
-
-  .actions {
-    display: flex;
-    justify-content: center;
-    gap: 0.5rem;
-
-    ion-button {
-      font-size: var(--ion-font-size-sm);
-      --padding-start: 0.5rem;
-      --padding-end: 0.5rem;
-    }
+/* ===============================
+   Responsivo
+================================== */
+@media (max-width: 768px) {
+  .actions-bar {
+    flex-direction: column;
+    align-items: stretch;
   }
 }
