+++ frontend/src/app/users/pages/users.page.ts
@@ -3,29 +3,44 @@ import { Router } from '@angular/router';
 import { CommonModule } from '@angular/common';
 import {
   IonContent,
-  IonHeader,
-  IonToolbar,
-  IonTitle,
   IonButton,
-  IonButtons,
   IonIcon,
-  IonGrid,
-  IonRow,
-  IonCol,
-} from '@ionic/angular';
+  IonSearchbar,
+} from '@ionic/angular/standalone';
+import { FormsModule } from '@angular/forms';
+import { AuthService } from '../../services/auth.service';
 import { AddUserModalComponent } from '../components/add-user/add-user-modal.component';
+import { HeaderComponent } from '../../shared/header/header.component';
 import { ConfirmModalComponent } from '../../shared/confirm-modal/confirm-modal.component';
-import { UserService, User } from '../services/user.service';
-import { AuthService } from '../../auth/services/auth.service';
+import { UserStatsComponent } from '../components/user-stats/user-stats.component';
+import { UsersTableComponent } from '../components/users-table/users-table.component';
+import { User, UserService } from '../services/user.service';
 import { UiService } from '../../core/services/ui.service';
 
 @Component({
   selector: 'app-users',
   templateUrl: './users.page.html',
   styleUrls: ['./users.page.scss'],
+  standalone: true,
+  imports: [
+    CommonModule,
+    FormsModule, // ✅ Necessário para ngModel funcionar
+    IonContent,
+    IonButton,
+    IonIcon,
+    IonSearchbar,
+    HeaderComponent,
+    AddUserModalComponent,
+    ConfirmModalComponent,
+    UserStatsComponent,
+    UsersTableComponent,
+  ],
 })
 export class UsersPage implements OnInit {
   users: User[] = [];
+  search: string = '';
+  currentUser: User | null = null;
+
   @ViewChild(AddUserModalComponent) addModal!: AddUserModalComponent;
   @ViewChild(ConfirmModalComponent) confirmModal!: ConfirmModalComponent;
 
@@ -33,19 +48,41 @@ export class UsersPage implements OnInit {
     private userService: UserService,
     private auth: AuthService,
     private router: Router,
-    private ui: UiService,
+    private ui: UiService
   ) {}
 
   ngOnInit() {
-    this.load();
+    if (this.auth.isLoggedIn()) {
+      this.load();
+    } else {
+      this.router.navigateByUrl('/login');
+    }
   }
 
   ionViewWillEnter() {
-    this.load();
+    if (this.auth.isLoggedIn()) {
+      this.load();
+    } else {
+      this.router.navigateByUrl('/login');
+    }
   }
 
   async load() {
     this.users = await this.userService.findAll();
+
+    // get the userId from the token and resolve the current user
+    const id = this.auth.userId;
+    this.currentUser = id ? this.users.find((u) => u.id === id) ?? null : null;
+  }
+
+  filteredUsers(): User[] {
+    if (!this.search.trim()) return this.users;
+    const query = this.search.toLowerCase();
+    return this.users.filter(
+      (u) =>
+        u.name.toLowerCase().includes(query) ||
+        u.email.toLowerCase().includes(query)
+    );
   }
 
   addUser() {
@@ -68,24 +105,24 @@ export class UsersPage implements OnInit {
   }
 
   async deleteUser(user: User) {
-    const ok = await this.confirmModal.open('Delete user?');
+    const ok = await this.confirmModal.open('Deseja excluir este usuário?');
     if (!ok) return;
     try {
       await this.userService.delete(user.id);
-      this.ui.toast('User deleted', 'success');
+      this.ui.toast('Usuário excluído', 'success');
       if (user.id === this.auth.userId) {
         this.logout();
       } else {
         this.load();
       }
     } catch {
-      // error handled globally
+      // handled globally
     }
   }
 
   logout() {
     this.auth.logout();
     this.router.navigateByUrl('/login');
-    this.ui.toast('Logged out', 'success');
+    this.ui.toast('Sessão encerrada', 'success');
   }
 }
