+++ frontend/src/app/users/users-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { UsersPage } from './users.page';
+import { UsersPage } from './pages/users.page';
 
 const routes: Routes = [{ path: '', component: UsersPage }];
