+++ frontend/src/app/users/users.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { IonicModule } from '@ionic/angular';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { UsersRoutingModule } from './users-routing.module';
-import { UsersPage } from './users.page';
+import { UsersPage } from './pages/users.page';
 import { AddUserModalComponent } from './components/add-user/add-user-modal.component';
 import { ConfirmModalComponent } from '../shared/confirm-modal/confirm-modal.component';
