--- frontend/src/app/users/users.page.html
@@ -1,43 +1,0 @@
-<ion-content class="users-wrapper" fullscreen>
-  <app-header></app-header>
-  <div class="users-container">
-    <!-- Cabeçalho -->
-    <div class="header-section">
-      <h1 class="text-title-xl">Gerenciamento de Usuários</h1>
-      <p class="text-muted">Sistema de autenticação seguro com JWT</p>
-    </div>
-
-    <!-- Cards de Resumo -->
-    <app-user-stats [total]="users.length" [currentUser]="currentUser"></app-user-stats>
-
-    <!-- Ações -->
-    <div class="actions-bar">
-      <ion-searchbar
-        [(ngModel)]="search"
-        placeholder="Buscar por nome ou email..."
-        class="user-search"
-      ></ion-searchbar>
-
-      <ion-button (click)="addUser()" class="btn-primary" shape="round">
-        <ion-icon slot="start" name="person-add-outline"></ion-icon>
-        Novo Usuário
-      </ion-button>
-    </div>
-
-    <!-- Tabela -->
-    <app-users-table
-      [users]="filteredUsers()"
-      [currentUser]="currentUser"
-      (edit)="editUser($event)"
-      (delete)="deleteUser($event)"
-    ></app-users-table>
-
-    <!-- Modais -->
-    <app-add-user-modal
-      #addModal
-      (userCreated)="userAdded($event)"
-      (userUpdated)="userUpdated($event)"
-    ></app-add-user-modal>
-    <app-confirm-modal #confirmModal></app-confirm-modal>
-  </div>
-</ion-content>
