--- frontend/src/app/users/users.page.scss
@@ -1,72 +1,0 @@
-.users-wrapper {
-  background: var(--ion-background-color);
-  padding: 2rem 1rem 3rem;
-}
-
-.users-container {
-  max-width: 1100px;
-  padding: 0 1rem;
-  margin: 0 auto;
-  display: flex;
-  flex-direction: column;
-  gap: 2.5rem;
-}
-
-/* ===============================
-   Cabeçalho
-================================== */
-.header-section {
-  text-align: center;
-
-  h1 {
-    font-size: 1.75rem;
-    font-weight: 700;
-    color: #1f2937;
-    margin-bottom: 0.5rem;
-  }
-
-  p {
-    font-size: 0.95rem;
-    color: #6b7280;
-  }
-}
-
-
-/* ===============================
-   Barra de ações
-================================== */
-.actions-bar {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 1rem;
-  justify-content: space-between;
-  align-items: center;
-
-  .user-search {
-    flex: 1;
-    padding: 0;
-    --background: white;
-    --border-radius: 12px;
-    border-radius: 12px;
-    box-shadow: 0 1px 2px rgba(0, 0, 0, 0.05);
-  }
-
-  ion-button {
-    height: 44px;
-    font-weight: 600;
-    --background: var(--ion-color-primary);
-    --border-radius: 12px;
-    --color: var(--ion-color-primary-contrast);
-  }
-}
-
-
-/* ===============================
-   Responsivo
-================================== */
-@media (max-width: 768px) {
-  .actions-bar {
-    flex-direction: column;
-    align-items: stretch;
-  }
-}
